Add test for App and Button Component
