+++ package.json
@@ -12,6 +12,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.1",
     "react-scripts": "4.0.3",
+    "react-test-renderer": "^17.0.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -43,7 +44,6 @@
     "@babel/eslint-parser": "^7.15.0",
     "@babel/plugin-syntax-jsx": "^7.14.5",
     "@babel/preset-react": "^7.14.5",
-    "babel-jest": "^27.1.0",
     "babel-preset-stage-0": "^6.24.1",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
@@ -53,9 +53,7 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.25.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "jest-cli": "^27.1.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-test-renderer": "^17.0.2",
     "redux-mock-store": "^1.5.4",
     "sinon": "^11.1.2",
     "stylelint": "^13.13.1",
