+++ src/ __tests__/components/App.test.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import App from '../../components/App';
+
+describe('App', () => {
+  it('should render calculator', () => {
+    const tree = renderer
+      .create(<App />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
